Dash - Break out logic into modules.
Start creating logic pieces and organize them into modules.
